+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## NEXT
 
 * Add build on node 14.x & issue templates
+* Add "Invert Comfort Level" setting (for Evaporative Cooling)
 
 ## 3.2.1 (2020-11-07)